examples refactoring
